Added ToDoTask to the Project interface and created ToDoTask inteface
