+++ src/components/EditProjectForm.tsx
@@ -1,5 +1,5 @@
 import React, {FormEvent, useState} from 'react'
-import { EditProjectFormProps } from '../types'
+import { Project } from '../types'
 
 
 
@@ -23,3 +23,7 @@ export const EditProjectForm: React.FC<EditProjectFormProps> = ({editProject, pr
     )
 }
 
+export interface EditProjectFormProps {
+  editProject: (projectName: string, id:string) => void;
+  project: Project;
+}
